# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ServiceDelivery.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/job-offers/{id}/accept": {
      "post": {
        "tags": [
          "JobOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptJobOfferResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptJobOfferResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptJobOfferResult"
                }
              }
            }
          }
        }
      }
    },
    "/job-offers/{id}/decline": {
      "post": {
        "tags": [
          "JobOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeclineJobOfferResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeclineJobOfferResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeclineJobOfferResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/service-requests/{id}": {
      "get": {
        "tags": [
          "ServiceRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/service-requests": {
      "get": {
        "tags": [
          "ServiceRequests"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveServiceRequestDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveServiceRequestDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveServiceRequestDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitServiceRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitServiceRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitServiceRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitServiceRequestResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitServiceRequestResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitServiceRequestResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/vehicles/{id}/claim": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimVehicleResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimVehicleResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimVehicleResult"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/force-release": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ForceReleaseVehicleResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForceReleaseVehicleResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForceReleaseVehicleResult"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/release": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseVehicleResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseVehicleResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseVehicleResult"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/take-over": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TakeOverVehicleResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakeOverVehicleResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakeOverVehicleResult"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}/position": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehiclePositionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehiclePositionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehiclePositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateVehiclePositionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateVehiclePositionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateVehiclePositionResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcceptJobOfferResult": {
        "required": [
          "offerId",
          "requestId",
          "offerStatus",
          "requestStatus",
          "repState"
        ],
        "type": "object",
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "offerStatus": {
            "type": "string"
          },
          "requestStatus": {
            "type": "string"
          },
          "repState": {
            "type": "string"
          }
        }
      },
...
      "ActiveServiceRequestDto": {
        "required": [
          "requestId",
          "requesterName",
          "tier",
          "dtcTitle",
          "status",
          "assignedRepId",
          "assignedRepName",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requesterName": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "dtcTitle": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "assignedRepId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "assignedRepName": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AssignedRepDto": {
        "required": [
          "repId",
          "name"
        ],
        "type": "object",
        "properties": {
          "repId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ClaimVehicleResult": {
        "required": [
          "vehicleId",
          "repId",
          "claimedAt"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "repId": {
            "type": "string",
            "format": "uuid"
          },
          "claimedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DeclineJobOfferResult": {
        "required": [
          "offerId",
          "requestId",
          "offerStatus"
        ],
        "type": "object",
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "offerStatus": {
            "type": "string"
          }
        }
      },
...
      "ForceReleaseVehicleResult": {
        "required": [
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "JobOfferHistoryDto": {
        "required": [
          "offerId",
          "repId",
          "repName",
          "status",
          "offeredAt",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "repId": {
            "type": "string",
            "format": "uuid"
          },
          "repName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "offeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LocationDto": {
        "required": [
          "lat",
          "lng"
        ],
        "type": "object",
        "properties": {
          "lat": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lng": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "ReleaseVehicleResult": {
        "required": [
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ServiceRequestDetailDto": {
        "required": [
          "requestId",
          "requesterName",
          "tier",
          "dtcTitle",
          "requesterLocation",
          "status",
          "assignedRep",
          "createdAt",
          "offerHistory"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requesterName": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "dtcTitle": {
            "type": "string"
          },
          "requesterLocation": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "status": {
            "type": "string"
          },
          "assignedRep": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AssignedRepDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "offerHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobOfferHistoryDto"
            }
          }
        }
      },
...
      "SubmitServiceRequestBody": {
        "required": [
          "dtcId",
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "dtcId": {
            "type": "string",
            "format": "uuid"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "SubmitServiceRequestResult": {
        "required": [
          "requestId",
          "status"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "TakeOverVehicleResult": {
        "required": [
          "sessionId",
          "vehicleId",
          "repId",
          "repState",
          "startedAt"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "repId": {
            "type": "string",
            "format": "uuid"
          },
          "repState": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateVehiclePositionRequest": {
        "required": [
          "latitude",
          "longitude",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateVehiclePositionResult": {
        "required": [
          "vehicleId"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}