# src/ServiceDelivery.Api/Controllers/VehiclesController.cs
using System.Security.Claims;
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.EntityFrameworkCore;
using Microsoft.AspNetCore.Mvc;
using ServiceDelivery.Application.Features.Vehicles.Commands;
using ServiceDelivery.Application.Features.Vehicles.Queries;
using ServiceDelivery.Domain.Exceptions;

namespace ServiceDelivery.Api.Controllers;

public record UpdateVehiclePositionRequest(double Latitude, double Longitude, DateTime Timestamp);

[ApiController]
[Route("vehicles")]
[Authorize]
public class VehiclesController : ControllerBase
{
    private readonly IMediator _mediator;

    public VehiclesController(IMediator mediator)
    {
        _mediator = mediator;
    }

    [HttpGet]
    [Authorize(Roles = "Dispatcher")]
    [ProducesResponseType<IReadOnlyList<VehicleDto>>(StatusCodes.Status200OK)]
    public async Task<IActionResult> GetFleet()
    {
        var dealerIdClaim = User.FindFirstValue("dealerId");

        if (!Guid.TryParse(dealerIdClaim, out var dealerId))
            return Unauthorized();

        var result = await _mediator.Send(new GetFleetQuery(dealerId));
        return Ok(result);
    }

    [HttpGet("available")]
    [Authorize(Roles = "ServiceRep")]
    [ProducesResponseType<IReadOnlyList<AvailableVehicleDto>>(StatusCodes.Status200OK)]
    public async Task<IActionResult> GetAvailableVehicles()
    {
        var dealerIdClaim = User.FindFirstValue("dealerId");

        if (!Guid.TryParse(dealerIdClaim, out var dealerId))
            return Unauthorized();

        var result = await _mediator.Send(new GetAvailableVehiclesQuery(dealerId));
        return Ok(result);
    }

    [HttpPost("{id:guid}/claim")]
    [Authorize(Roles = "ServiceRep")]
    [ProducesResponseType<ClaimVehicleResult>(StatusCodes.Status200OK)]
    public async Task<IActionResult> ClaimVehicle(Guid id)
    {
        var repIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);

        if (!Guid.TryParse(repIdClaim, out var repId))
            return Unauthorized();

        try
        {
            var result = await _mediator.Send(new ClaimVehicleCommand(id, repId));
            return Ok(result);
        }
        catch (KeyNotFoundException)
        {
            return NotFound();
        }
        catch (VehicleAlreadyClaimedException ex)
        {
            return Conflict(new { error = ex.Message });
        }
        catch (RepAlreadyHasActiveSessionException ex)
        {
            return Conflict(new { error = ex.Message });
        }
    }

    [HttpPost("{id:guid}/force-release")]
    [Authorize(Roles = "Dispatcher")]
    [ProducesResponseType<ForceReleaseVehicleResult>(StatusCodes.Status200OK)]
    public async Task<IActionResult> ForceReleaseVehicle(Guid id)
    {
        try
        {
            var result = await _mediator.Send(new ForceReleaseVehicleCommand(id));
            return Ok(result);
        }
        catch (KeyNotFoundException)
        {
            return NotFound();
        }
    }

    [HttpPost("{id:guid}/release")]
    [Authorize(Roles = "ServiceRep")]
    [ProducesResponseType<ReleaseVehicleResult>(StatusCodes.Status200OK)]
    public async Task<IActionResult> ReleaseVehicle(Guid id)
    {
        var repIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);

        if (!Guid.TryParse(repIdClaim, out var repId))
            return Unauthorized();

        try
        {
            var result = await _mediator.Send(new ReleaseVehicleCommand(id, repId));
            return Ok(result);
        }
        catch (KeyNotFoundException)
        {
            return NotFound();
        }
        catch (VehicleNotClaimedByRepException ex)
        {
            return BadRequest(new { error = ex.Message });
        }
        catch (VehicleReleaseBlockedByActiveJobException ex)
        {
            return BadRequest(new { error = ex.Message });
        }
    }

    [HttpPost("{id:guid}/take-over")]
    [Authorize(Roles = "ServiceRep")]
    [ProducesResponseType<TakeOverVehicleResult>(StatusCodes.Status200OK)]
    public async Task<IActionResult> TakeOver(Guid id)
    {
        var repIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);

        if (!Guid.TryParse(repIdClaim, out var repId))
            return Unauthorized();

        try
        {
            var result = await _mediator.Send(new TakeOverVehicleCommand(id, repId));
            return Ok(result);
        }
        catch (KeyNotFoundException)
        {
            return NotFound();
        }
        catch (RepNotIdleException ex)
        {
            return Conflict(new { reason = ex.Message });
        }
        catch (VehicleNotIdleException ex)
        {
            return Conflict(new { reason = ex.Message });
        }
        catch (DbUpdateConcurrencyException)
        {
            return Conflict(new { reason = "Vehicle is no longer idle; it was taken over concurrently." });
        }
    }

    [HttpPost("{id:guid}/position")]
    [Authorize(Roles = "Simulator")]
    [ProducesResponseType<UpdateVehiclePositionResult>(StatusCodes.Status200OK)]
    public async Task<IActionResult> UpdateVehiclePosition(Guid id, [FromBody] UpdateVehiclePositionRequest body)
    {
        var simulatorUserIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);

        if (!Guid.TryParse(simulatorUserIdClaim, out var simulatorUserId))
            return Unauthorized();

        try
        {
            var result = await _mediator.Send(new UpdateVehiclePositionCommand(
                id, simulatorUserId, body.Latitude, body.Longitude, body.Timestamp));
            return Ok(result);
        }
        catch (KeyNotFoundException)
        {
            return NotFound();
        }
    }
}